Reorganize header files
* random_64bit_generator and random_64bit_accessor w/o the constructor definition go to dqrng_types.h
* this gets automatically included in dqrng_RcppExports.h.
* constructor definition for random_64bit_accessor goes to dqrng.h.
* boost modifications go to dqrng_distributions.h
