+++ src/RcppExports.cpp
@@ -2,6 +2,7 @@
 // Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
 #include "../inst/include/dqrng.h"
+#include "../inst/include/dqrng_types.h"
 #include <Rcpp.h>
 #include <string>
 #include <set>
